+++ src/it/test-projects/creedengo-java-plugin-test-project/src/main/java/org/greencodeinitiative/creedengo/java/checks/GCI82/MakeNonReassignedVariablesConstantsForRecord.java
@@ -1,0 +1,9 @@
+package org.greencodeinitiative.creedengo.java.checks;
+
+public class MakeNonReassignedVariablesConstantsForRecord {
+
+    private record myRecord(
+            String myImplicitlyFinalStringField, // Compliant
+            Integer myImplicitlyFinalIntField) // Compliant
+    { }
+}